Implementada Validacion Usuarios

+++ src/main/java/com/javi/tareas/controllers/TaskController.java
@@ -1,5 +1,6 @@
 package com.javi.tareas.controllers;
 
+import com.javi.tareas.component.SessionValidator;
 import com.javi.tareas.entities.Status;
 import com.javi.tareas.entities.Task;
 import com.javi.tareas.services.TaskServices;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import java.util.List;
 
 
 /**
@@ -23,6 +25,10 @@ import java.time.LocalTime;
 @Slf4j // Anotación que habilita la funcionalidad de registro (logging) en la clase.
 @Controller // Marca esta clase como un controlador de Spring MVC.
 public class TaskController {
+
+    @Autowired
+    private SessionValidator sessionValidator;
+
     @Autowired // Inyecta una instancia del bean TaskServices
     private TaskServices taskService;
 
@@ -38,7 +44,7 @@ public class TaskController {
      * Método privado que se encarga de ordenar una lista de tareas y agregarla a un modelo, según la opción de ordenación especificada.
      *
      * @param sortOption La opción de ordenación que determina cómo se ordenarán las tareas.
-     * @param model El modelo al que se añadirá la lista de tareas ordenadas.
+     * @param model      El modelo al que se añadirá la lista de tareas ordenadas.
      */
     private void sortAndAddToModel(String sortOption, Model model) {
         switch (sortOption) {
@@ -52,20 +58,33 @@ public class TaskController {
     /**
      * Método controlador que se encarga de mostrar la página principal de tareas para un usuario específico.
      *
-     * @param id El ID del usuario para el cual se mostrarán las tareas.
+     * @param id         El ID del usuario para el cual se mostrarán las tareas.
      * @param sortOption Cookie que almacena la opción de ordenación
-     * @param model El modelo utilizado para pasar datos a la vista.
+     * @param model      El modelo utilizado para pasar datos a la vista.
      * @return La vista "index" que muestra las tareas del usuario.
      */
     @GetMapping("/home/{id}")
-    public String home(@PathVariable("id") Long id, @CookieValue(name = "sortOption", defaultValue = "status") String sortOption ,Model model) {
-        // Establece el ID del usuario actual
-        userId = id;
+    public String home(@PathVariable("id") Long id, @CookieValue(name = "sortOption", defaultValue = "status") String sortOption
+                       ,@RequestParam(name = "search", required = false) String searchTitle
+                        , HttpSession session
+                       , Model model) {
+        // Establece el ID del usuario actual y lo guardamos en una cookie
 
-        // Ordenamos las tareas y las añadimos al modelo
-        sortAndAddToModel(sortOption, model);
-        model.addAttribute("sortOption", sortOption);
-        return "index";
+        if (sessionValidator.isValidUserSession(session, id)) {
+            userId = id;
+            if (searchTitle == null) {
+                sortAndAddToModel(sortOption, model);
+                model.addAttribute("sortOption", sortOption);
+
+            } else {
+                List<Task> taskFind = taskService.findTask(searchTitle, userId);
+                model.addAttribute("taskList", taskFind);
+            }
+            return "/taskHome/index";
+        } else {
+            return "/logIn/index";
+        }
+
     }
 
     /**
@@ -79,14 +98,18 @@ public class TaskController {
      * @return La vista "new-task" que muestra un formulario para añadir la nueva tarea.
      */
     @GetMapping("/newTask")
-    public String newTask(Model model) {
-        Task task = Task.builder()
-                .dueDate(LocalDate.now())
-                .time(LocalTime.now())
-                .idUser(userId)
-                .build();
-        model.addAttribute("taskDt", task);
-        return "new-task";
+    public String newTask(HttpSession session ,Model model) {
+        if (sessionValidator.isValidUserSession(session, userId)) {
+
+            Task task = Task.builder()
+                    .dueDate(LocalDate.now())
+                    .time(LocalTime.now())
+                    .idUser(userId)
+                    .build();
+            model.addAttribute("taskDt", task);
+            return "taskHome/new-task";
+        }
+        return "logIn/index";
     }
 
     /**
@@ -130,9 +153,12 @@ public class TaskController {
      * @return La vista "task" que muestra los detalles de la tarea.
      */
     @GetMapping("/viewTask")
-    public String viewTask(@RequestParam("id") Long id, Model model) {
-        model.addAttribute("taskDt", taskService.find(id));
-        return "task";
+    public String viewTask(@RequestParam("id") Long id,HttpSession session, Model model) {
+        if (sessionValidator.isValidUserSession(session, userId)) {
+            model.addAttribute("taskDt", taskService.find(id));
+            return "taskHome/task";
+        }
+        return "logIn/index";
     }
 
     /**
@@ -142,9 +168,14 @@ public class TaskController {
      * @return Redirige al usuario a su página de inicio después de eliminar la tarea.
      */
     @GetMapping("/deleteTask")
-    public String delete(@RequestParam("task") Long idTask) {
-        taskService.delete(idTask);
-        return "redirect:/home/" + userId;
+    public String delete(@RequestParam("task") Long idTask, HttpSession session) {
+
+        if (sessionValidator.isValidUserSession(session, userId)) {
+            taskService.delete(idTask);
+            return "redirect:/home/" + userId;
+        } else {
+            return "logIn/index";
+        }
     }
 
     /**
@@ -155,10 +186,14 @@ public class TaskController {
      * @return La vista "edit-task", un formulario con los detalles de la tarea que se desea editar.
      */
     @GetMapping("/editTask")
-    public String edit(@RequestParam("idTask") Long idTask, Model model) {
-        Task t = taskService.find(idTask);
-        model.addAttribute("taskDt", t);
-        return "edit-task";
+    public String edit(@RequestParam("idTask") Long idTask, HttpSession session, Model model) {
+        if (sessionValidator.isValidUserSession(session, userId)) {
+            Task t = taskService.find(idTask);
+            model.addAttribute("taskDt", t);
+            return "taskHome/edit-task";
+        } else {
+            return "logIn/index";
+        }
     }
 
     /**
@@ -170,23 +205,23 @@ public class TaskController {
      * - Si la validación es exitosa, actualiza la tarea en el repositorio y redirige al usuario a la vista de detalles de la tarea.
      *
      * @param editTask La tarea editada proporcionada a través del formulario y anotada como válida.
-     * @param result    El resultado de la validación que se utilizará para detectar errores.
+     * @param result   El resultado de la validación que se utilizará para detectar errores.
      * @return Si hay errores de validación, retorna la vista "edit-task". Si la tarea se actualiza exitosamente,
      * redirige al usuario a la vista de detalles de la tarea.
      */
     @PostMapping("/editTask/submit")
     public String updateTask(@Valid @ModelAttribute("taskDt") Task editTask, BindingResult result) {
         if (result.hasErrors()) {
-            return "edit-task";
+            return "taskHome/edit-task";
         } else {
             if (taskService.timeNullValid(editTask)) {
                 result.rejectValue("time", "time.error");
-                return "edit-task";
+                return "taskHome/edit-task";
             }
         }
 
         taskService.updateTask(editTask);
-        return "redirect:/viewTask/" + editTask.getId();
+        return "redirect:/viewTask?id=" + editTask.getId();
     }
 
     /**
@@ -206,7 +241,7 @@ public class TaskController {
      * Método controlador que permite al usuario ordenar las tareas en su página de inicio según una opción seleccionada.
      *
      * @param sortOption La opción de ordenación seleccionada por el usuario.
-     * @param response respuesta HTTP para generar una cookie con la elección de ordenación del usuario
+     * @param response   respuesta HTTP para generar una cookie con la elección de ordenación del usuario
      * @return Redirige al usuario a su página de inicio con las tareas ordenadas.
      */
     @GetMapping("/home/sortBy")
@@ -214,6 +249,7 @@ public class TaskController {
         if (sortOption == null) {
             sortOption = "status";
         }
+
 
         Cookie cookieSort = new Cookie("sortOption", sortOption);
         cookieSort.setMaxAge(COOKIE_MAX_RANGE);
@@ -223,5 +259,16 @@ public class TaskController {
 
         return "redirect:/home/" + userId;
     }
+
+    @GetMapping("/home/searchTask")
+    public String searchTask(@RequestParam("search") String title, HttpServletResponse response) {
+        Cookie cookieTitle = new Cookie("titleFilter", title);
+        cookieTitle.setMaxAge(-1);
+        cookieTitle.setPath("/home");
+
+        response.addCookie(cookieTitle);
+        return "redirect:/home/" + userId;
+    }
+
 
 }
